Refactor more stuff
