+++ test/suites/zombie_tanssi_keep_db/test_zombie_tanssi_keep_db.ts
@@ -8,7 +8,6 @@ import {
     directoryExists,
     findCollatorProcessPid,
     getAuthorFromDigest,
-    getAuthorFromDigestRange,
     getHeaderFromRelay,
     getKeyringNimbusIdHex,
     getTmpZombiePath,
